requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
google-auth==2.3.3
google-auth-oauthlib==0.4.6
gspread==4.0.1
httplib2==0.20.2
idna==3.3
oauth2client==4.1.3
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.5
PySocks==1.7.1
python-dotenv==0.19.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tweepy==3.10.0
urllib3==1.26.7